Выполнил заданМодуляцияÐ
