+++ js/utils/get-random-positive-float.js
@@ -21,4 +21,4 @@ function getRandomPositiveFloat (min, max, digits = 1) {
   return result.toFixed(digits);
 }
 
-
+export {getRandomPositiveFloat};
