+++ js/utils/get-random-positive-integer.js
@@ -26,3 +26,5 @@ function getRandomPositiveInteger (min, max) {
   // потому что Math.random() генерирует только дробные числа и ноль.
   return Math.floor(result);
 }
+
+export {getRandomPositiveInteger};
