+++ package.json
@@ -2,6 +2,7 @@
   "name": "keksobooking",
   "version": "23.0.0",
   "private": true,
+  "main": "js/main.js",
   "description": "Личный проект «Кексобукинг» от HTML Academy",
   "repository": {
     "type": "git",
@@ -13,7 +14,7 @@
   "devDependencies": {
     "browser-sync": "2.26.14",
     "eslint": "7.18.0",
-    "eslint-config-htmlacademy": "4.0.0"
+    "eslint-config-htmlacademy": "4.0.1"
   },
   "license": "MIT",
   "dependencies": {},
@@ -21,8 +22,10 @@
     "start": "browser-sync start --server --no-ui --files \"js/**/*.js\"",
     "test": "eslint js/"
   },
+  "type": "module",
   "engines": {
     "node": "14.x",
     "npm": "6.x"
   }
 }
+
